# apps/api/openapi.gen.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Char AI API",
    "description": "AI services API for speech-to-text transcription, LLM chat completions, and subscription management",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/nango/connections": {
      "get": {
        "tags": [
          "nango"
        ],
        "operationId": "list_connections",
        "responses": {
          "200": {
            "description": "List of active connections",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListConnectionsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "nango"
        ],
        "operationId": "delete_connection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Connection disconnected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteConnectionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/nango/session": {
      "post": {
        "tags": [
          "nango"
        ],
        "operationId": "create_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/subscription/can-start-trial": {
      "get": {
        "tags": [
          "subscription"
        ],
        "operationId": "can_start_trial",
        "responses": {
          "200": {
            "description": "Check successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CanStartTrialResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/subscription/start-trial": {
      "post": {
        "tags": [
          "subscription"
        ],
        "operationId": "start_trial",
        "parameters": [
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Interval"
            },
            "example": "monthly"
          }
        ],
        "responses": {
          "200": {
            "description": "Trial started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartTrialResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CanStartTrialReason": {
        "type": "string",
        "enum": [
          "eligible",
          "not_eligible",
          "error"
        ]
      },
      "CanStartTrialResponse": {
        "type": "object",
        "required": [
          "canStartTrial"
        ],
        "properties": {
          "canStartTrial": {
            "type": "boolean",
            "example": true
          },
          "reason": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CanStartTrialReason"
              }
            ]
          }
        }
      },
...
      "ConnectionItem": {
        "type": "object",
        "required": [
          "integration_id",
          "connection_id"
        ],
        "properties": {
          "connection_id": {
            "type": "string"
          },
          "integration_id": {
            "type": "string"
          },
          "last_error_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_error_description": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_error_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated_at": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "CreateSessionRequest": {
        "type": "object",
        "required": [
          "integration_id"
        ],
        "properties": {
          "integration_id": {
            "type": "string"
          }
        }
      },
...
      "DeleteConnectionRequest": {
        "type": "object",
        "required": [
          "connection_id",
          "integration_id"
        ],
        "properties": {
          "connection_id": {
            "type": "string"
          },
          "integration_id": {
            "type": "string"
          }
        }
      },
      "DeleteConnectionResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
      "Interval": {
        "type": "string",
        "enum": [
          "monthly",
          "yearly"
        ]
      },
...
      "ListConnectionsResponse": {
        "type": "object",
        "required": [
          "connections"
        ],
        "properties": {
          "connections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectionItem"
            }
          }
        }
      },
...
      "SessionResponse": {
        "type": "object",
        "required": [
          "token",
          "expires_at",
          "mode"
        ],
        "properties": {
          "connection_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "expires_at": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
      "StartTrialReason": {
        "type": "string",
        "enum": [
          "started",
          "not_eligible"
        ]
      },
      "StartTrialResponse": {
        "type": "object",
        "required": [
          "started"
        ],
        "properties": {
          "reason": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StartTrialReason"
              }
            ]
          },
          "started": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token"
      }
    }
  },
...
}